Refactor particle system into reusable widget
Moved particle system logic from InitialPage to a new reusable ParticleSystemWidget in utils/effects/particle_system.dart. Updated InitialPage to use the new widget and removed duplicate code. Also fixed import path casing in main.dart.
